feat: auth
- added an animated night sky with full moon (moonCode logo) svg to the auth pages (login && register)
- installed the `vite-plugin-svgr` to render svgs directly in the DOM
- grouped the login and register page components under an auth route group
