+++ apps/dashboard/src/components/layout/header/Logo.tsx
@@ -1,5 +1,11 @@
-const Logo = () => {
-  return <img src="/moon.svg" className="size-8" alt="MoonCode Logo" />;
-};
+import { cn } from "@repo/ui/lib/utils";
+
+const Logo = ({ className }: { className?: string }) => (
+  <img
+    src="/moon.svg"
+    className={cn("size-8", className)}
+    alt="MoonCode Logo"
+  />
+);
 
 export default Logo;
