+++ apps/dashboard/src/vite-env.d.ts
@@ -1,4 +1,6 @@
 /// <reference types="vite/client" />
+/// <reference types="vite-plugin-svgr/client" />
+
 interface ImportMetaEnv {
   readonly VITE_API_URL: string;
   readonly VITE_LOGIN_URL: string;
